Added 4 more questions
GetRangeBST- Given a Binary Search Tree and two numbers - a & b, return
all the nodes in the tree that lie
in the range [a .. b]. Your method should return an ArrayList with the
data of the qualifying nodes inserted
in ascending order.

MaxRepetition - Given an Array of integers, write a method that will
return the integer with the maximum number of repetitions. Your code is
expected to run with O(n) time complexity and O(1) space complexity.
The elements in the array are between 0 to size(array) - 1 and the
array will not be empty.

PrintPaths - ou're given a 2D board which contains an m x n matrix of
chars - char[][] board.
Write a method - printPaths that prints all possible paths from the top
left cell to the bottom right cell.
Your method should return an ArrayList of Strings, where each String
represents a path with characters appended in the order of movement.
You're only allowed to move down and right on the board.
The order of String insertion in the ArrayList does not matter.

ShiftK - Rotate an array to the left by k positions without using extra
space.k can be greater than the size of the array.