3 more questions
decompresstree - Decompress a string into a binary tree
pathLengthFromRoot - Given the root of a Binary Tree  and an integer
that represents the data value of a TreeNode present in the tree,
write a method - pathLengthFromRoot that returns the distance between
the root and that node. You can assume that the given key exists in the
tree.
The distance is defined as the minimum number of nodes that must be
traversed to reach the target node.
ShiftK - Rotate an array to the left by k positions without using extra
space.k can be greater than the size of the array.